refactor: further reorganize api libs
