+++ apps/ensindexer/src/api/handlers/resolution-api.ts
@@ -5,13 +5,13 @@ import {
 } from "@ensnode/ensnode-sdk";
 import { Hono } from "hono";
 
-import { errorResponse } from "@/api/lib/error-response";
+import { errorResponse } from "@/api/lib/handlers/error-response";
+import { validate } from "@/api/lib/handlers/validate";
 import { canAccelerateResolution } from "@/api/lib/indexing-status/can-accelerate-resolution";
 import { resolveForward } from "@/api/lib/resolution/forward-resolution";
 import { resolvePrimaryNames } from "@/api/lib/resolution/multichain-primary-name-resolution";
 import { resolveReverse } from "@/api/lib/resolution/reverse-resolution";
 import { captureTrace } from "@/api/lib/tracing/protocol-tracing";
-import { validate } from "@/api/lib/validate";
 import { simpleMemoized } from "@/lib/simple-memoized";
 import { routes } from "@ensnode/ensnode-sdk/internal";
