+++ apps/ensindexer/src/api/lib/handlers/error-response.ts
